pancake: first implementation of panseL4/panmicrokit

+++ build_sdk.py
@@ -677,18 +677,25 @@ def build_sel4(
     shutil.copy(invocations_all, dest)
     dest.chmod(0o744)
 
-    include_dir = sdk_dir / "board" / board.name / config.name / "include"
-    for source in ("kernel_Config", "libsel4", "libsel4/sel4_Config", "libsel4/autoconf"):
+    def copy_kernel_includes(target_dir, source_dir):
         source_dir = sel4_install_dir / source / "include"
         for p in source_dir.rglob("*"):
             if not p.is_file():
                 continue
             rel = p.relative_to(source_dir)
-            dest = include_dir / rel
+            dest = target_dir / rel
             dest.parent.mkdir(exist_ok=True, parents=True)
             dest.unlink(missing_ok=True)
             shutil.copy(p, dest)
             dest.chmod(0o744)
+
+    for source in ("kernel_Config", "libsel4", "libsel4/sel4_Config", "libsel4/autoconf"):
+        source_dir = sel4_install_dir / source / "include"
+        copy_kernel_includes(sdk_dir / "board" / board.name / config.name / "include", source_dir)
+        copy_kernel_includes(
+            sdk_dir / "board" / board.name / config.name / "include" / "pancake-c-interop",
+            source_dir
+        )
 
     if not board.arch.is_x86():
         # only non-x86 platforms have this file to describe memory regions
@@ -786,7 +793,7 @@ def build_lib_component(
 ) -> None:
     """Build a specific library component.
 
-    Right now this is just libmicrokit.a
+    Right now these are just libmicrokit.a and libmicrokitpnk.a
     """
     sel4_dir = sdk_dir / "board" / board.name / config.name
     build_dir = build_dir / board.name / config.name / component_name
@@ -815,10 +822,11 @@ def build_lib_component(
 
     link_script = Path(component_name) / "microkit.ld"
     dest = lib_dir / "microkit.ld"
-    dest.unlink(missing_ok=True)
-    shutil.copy(link_script, dest)
-    # Make output read-only
-    dest.chmod(0o744)
+    if link_script.exists():
+        dest.unlink(missing_ok=True)
+        shutil.copy(link_script, dest)
+        # Make output read-only
+        dest.chmod(0o744)
 
     include_dir = sdk_dir / "board" / board.name / config.name / "include"
     source_dir = Path(component_name) / "include"
@@ -826,6 +834,11 @@ def build_lib_component(
         if not p.is_file():
             continue
         rel = p.relative_to(source_dir)
+        # The Pancake components currently only support RISC-V
+        if not board.arch.is_riscv() and rel.parts[0] in ("panmicrokit", "pansel4"):
+            continue
+        if not board.arch.is_riscv() and rel.name == "panmicrokit.h":
+            continue
         dest = include_dir / rel
         dest.parent.mkdir(exist_ok=True, parents=True)
         dest.unlink(missing_ok=True)
@@ -1006,6 +1019,9 @@ def main() -> None:
 
                 build_elf_component("monitor", sdk_dir, build_dir, board, config, args.llvm, [])
                 build_lib_component("libmicrokit", sdk_dir, build_dir, board, config, args.llvm)
+                if board.arch.is_riscv():
+                    # the Pancake version is RISCV64 only for now
+                    build_lib_component("libmicrokitpnk", sdk_dir, build_dir, board, config, args.llvm)
                 if not args.skip_initialiser:
                     build_initialiser("initialiser", sdk_dir, build_dir, board, config)
